Removed check for NULL pointer address in tempSensor - not possible with dynaconfig.
Sensor calibration adjustement.

reworked sensor connect/disconnect functionality.
I used to get repeated messages about the sensors being disconnected. This is fixed. Also a disconnected sensor doesn't hang the system now, since DallasTemperature::isConnected is used as an early test to see if a sensor with the address is present.