Fixes for OneWire actuator and sensor (DS2413)
pioMask() function was not returning correct value as was incrementing
after returning. Changed to increment before returning.

OneWireActuator sense() function was not applying invert option
correctly. This has been tested but there is a current issue with the
multiple inheritance of the "OneWireActuator" class for both "Actuator"
and SwitchSensor". If the order of these 2 are switched the sensor
functionality via DS2413 works correctly but actuators via DS2413 no
longer work. I have not been able to resolve this issue so for now door
sensor does not work via DS2413.