valve controller and onewire actuator share DS24xx driver
The OneWire actuators used to read the hardware, change some bits and write it back. The software did not know the state of the latch.
This commit changes this to having a DS24xx driver with caching, which both actuators have a reference too. When installing an actuator, it will search whether the IC is already used elsewhere and copy the shared_ptr from that object.