Pass lookup as const ref from base class instead of std::function by value
To be able to keep a lookup object outside of the library classes and to change the lookup data, it should not be passed by value. I changed it to a const ref, which is passed on construction. The lookup object should be kept, but the internal data can be changed, for example by cbox.