8 hours of debugging to find a missing *. Aren't void pointers just lovely
This missing * would cause the address of the OneWire bus to be slightly off. Which caused the OneWire methods to execute with the wrong this pointer. So when writing it's member variables, it actually overwrote the pb_ostream object. So the callback address was invalid when used, and this resulted in hardfault during protobuf encoding.